Renamed endian.h to endianapple.h to avoid collision, refactored build.sh to be relative with more options, added cleanup and unset environment scripts
